requirements.txt
#
# This file is autogenerated by pip-compile
# Make changes in requirements.in, then run this to update:
#
#    pip-compile requirements.in
#
alembic==0.8.4            # via dataset
arrow==0.7.0
click==6.2
dataset==0.6.2
Mako==1.0.3               # via alembic
MarkupSafe==0.23          # via mako
normality==0.2.4          # via dataset
path.py==8.1.2
python-dateutil==2.4.2
python-editor==0.5        # via alembic
PyYAML==3.11              # via dataset
simplejson==3.8.2
six==1.10.0
sqlalchemy==1.0.12        # via alembic, dataset
tablib==0.11.2
wheel==0.29.0
